feat: refactor summarization logic out of knowledge graph
