+++ src/KnowledgeGraph.ts
@@ -1,7 +1,6 @@
 import fs from 'node:fs/promises';
 import { v4 as uuid } from 'uuid';
 import { CFG } from './config.ts';
-import { SummarizationAgent } from './agents/summarize_agent.ts';
 
 // -----------------------------------------------------------------------------
 // Minimal JSON‑file Knowledge Graph adapter ------------------------------------
@@ -60,18 +59,9 @@ export interface SummarizationResult {
 export class KnowledgeGraphManager {
   public static WINDOW = CFG.WINDOW;
 
-  // Number of nodes after which to trigger summarization
-  private static SUMMARIZATION_THRESHOLD = 20;
-
-  // Maximum number of nodes to include in a summary
-  private static SUMMARY_CHUNK_SIZE = 10;
-
   private entities: Record<string, DagNode | ArtifactRef> = {};
   private relations: { from: string; to: string; type: string }[] = [];
   private dirty = false;
-
-  // Summarization agent instance
-  private summarizationAgent: SummarizationAgent | null = null;
 
   public labelIndex: Map<string, string> = new Map(); // branchLabel ➜ nodeId
 
@@ -190,194 +180,6 @@ export class KnowledgeGraphManager {
       parents: n.parents,
       artifacts: n.artifacts?.map((a) => ({ name: a.name, uri: a.uri })),
     }));
-  }
-
-  /**
-   * Gets or creates the summarization agent.
-   */
-  private getSummarizationAgent(): SummarizationAgent {
-    if (!this.summarizationAgent) {
-      this.summarizationAgent = new SummarizationAgent();
-    }
-    return this.summarizationAgent;
-  }
-
-  /**
-   * Checks if summarization is needed and triggers it if necessary.
-   * This should be called after adding new nodes to the graph.
-   */
-  async checkAndTriggerSummarization(): Promise<void> {
-    const branches = this.listBranches();
-
-    for (const branch of branches) {
-      // Only summarize branches that have enough nodes
-      if (branch.depth >= KnowledgeGraphManager.SUMMARIZATION_THRESHOLD) {
-        await this.summarizeBranch(branch.branchId);
-      }
-    }
-  }
-
-  /**
-   * Summarizes the oldest segment of nodes in a branch.
-   * @param branchId ID of the branch to summarize
-   * @returns A SummarizationResult object containing the summary or error information
-   */
-  async summarizeBranch(branchId: string): Promise<SummarizationResult> {
-    // Get the branch head
-    const head = this.getNode(branchId);
-    if (!head) {
-      console.error(`[summarizeBranch] Branch not found: ${branchId}`);
-      return {
-        summary: null,
-        success: false,
-        errorCode: 'BRANCH_NOT_FOUND',
-        errorMessage: `Branch with ID or label "${branchId}" not found`,
-      };
-    }
-
-    // Collect all nodes in the branch
-    const branchNodes: DagNode[] = [];
-    let current: DagNode | undefined = head;
-
-    while (current) {
-      branchNodes.push(current);
-      current = current.parents[0] ? this.getNode(current.parents[0]) : undefined;
-    }
-
-    // Reverse to get chronological order
-    branchNodes.reverse();
-
-    // Log branch information for debugging
-    console.log(`[summarizeBranch] Branch ${branchId} has ${branchNodes.length} nodes`);
-
-    // Check if the branch has enough nodes to meet the summarization threshold
-    if (branchNodes.length < KnowledgeGraphManager.SUMMARIZATION_THRESHOLD) {
-      return {
-        summary: null,
-        success: false,
-        errorCode: 'INSUFFICIENT_NODES',
-        errorMessage: `Branch has only ${branchNodes.length} nodes, which is below the summarization threshold of ${KnowledgeGraphManager.SUMMARIZATION_THRESHOLD}`,
-        details: `Current nodes: ${branchNodes.length}, Required: ${KnowledgeGraphManager.SUMMARIZATION_THRESHOLD}`,
-      };
-    }
-
-    // Check if we already have summaries for this branch
-    const existingSummaries = branchNodes.filter(
-      (node): node is SummaryNode =>
-        node.role === 'summary' && node.summarizedSegment !== undefined,
-    );
-
-    console.log(`[summarizeBranch] Branch has ${existingSummaries.length} existing summaries`);
-
-    // Determine which nodes need to be summarized
-    let nodesToSummarize: DagNode[] = [];
-
-    if (existingSummaries.length === 0) {
-      // If no summaries exist, summarize the oldest chunk
-      nodesToSummarize = branchNodes.slice(
-        0,
-        Math.min(KnowledgeGraphManager.SUMMARY_CHUNK_SIZE, branchNodes.length),
-      );
-    } else {
-      // Find the newest summary
-      const newestSummary = existingSummaries.reduce((newest, current) => {
-        const newestDate = new Date(newest.createdAt);
-        const currentDate = new Date(current.createdAt);
-        return currentDate > newestDate ? current : newest;
-      }, existingSummaries[0]);
-
-      // Find nodes that were created after the newest summary but are old enough to summarize
-      const summaryIndex = branchNodes.findIndex((node) => node.id === newestSummary.id);
-
-      if (
-        summaryIndex !== -1 &&
-        branchNodes.length - summaryIndex > KnowledgeGraphManager.SUMMARIZATION_THRESHOLD
-      ) {
-        nodesToSummarize = branchNodes.slice(
-          summaryIndex + 1,
-          summaryIndex + 1 + KnowledgeGraphManager.SUMMARY_CHUNK_SIZE,
-        );
-      }
-    }
-
-    // If there are nodes to summarize, create a summary
-    if (nodesToSummarize.length > 0) {
-      console.log(`[summarizeBranch] Summarizing ${nodesToSummarize.length} nodes`);
-      try {
-        const summaryNode = await this.createSummary(nodesToSummarize);
-        return {
-          summary: summaryNode,
-          success: true,
-        };
-      } catch (error) {
-        console.error(`[summarizeBranch] Error creating summary:`, error);
-        return {
-          summary: null,
-          success: false,
-          errorCode: 'SUMMARIZATION_ERROR',
-          errorMessage: 'Error occurred during summarization process',
-          details: error instanceof Error ? error.message : String(error),
-        };
-      }
-    }
-
-    // If we reach here, all nodes have already been summarized
-    return {
-      summary: null,
-      success: false,
-      errorCode: 'ALREADY_SUMMARIZED',
-      errorMessage: 'All nodes in this branch have already been summarized',
-      details: `Branch has ${existingSummaries.length} summaries covering all summarizable nodes`,
-    };
-  }
-
-  /**
-   * Creates a summary for a segment of nodes.
-   * @param nodes Nodes to summarize
-   * @throws Error if summarization fails
-   */
-  async createSummary(nodes: DagNode[]): Promise<SummaryNode> {
-    if (!nodes || nodes.length === 0) {
-      throw new Error('Cannot create summary: No nodes provided');
-    }
-
-    console.log(`[createSummary] Creating summary for ${nodes.length} nodes`);
-
-    const agent = this.getSummarizationAgent();
-    const result = await agent.summarize(nodes);
-
-    // Check for errors in the summarization result
-    if (result.error) {
-      console.error(`[createSummary] Summarization agent error:`, result.error);
-      throw new Error(`Summarization failed: ${result.error}`);
-    }
-
-    // Validate the summary content
-    if (!result.summary || result.summary.trim() === '') {
-      console.error(`[createSummary] Summarization agent returned empty summary`);
-      throw new Error('Summarization failed: Empty summary returned');
-    }
-
-    // Create a summary node
-    const summaryNode: SummaryNode = {
-      id: uuid(),
-      thought: result.summary,
-      role: 'summary',
-      parents: [nodes[nodes.length - 1].id], // Link to the newest node in the segment
-      children: [],
-      createdAt: new Date().toISOString(),
-      summarizedSegment: nodes.map((node) => node.id),
-      tags: ['summary'],
-    };
-
-    console.log(`[createSummary] Created summary node with ID ${summaryNode.id}`);
-
-    // Persist the summary node
-    this.createEntity(summaryNode);
-    this.createRelation(nodes[nodes.length - 1].id, summaryNode.id, 'has_summary');
-    await this.flush();
-
-    return summaryNode;
   }
 
   /* ------------------------- helpers ------------------------------- */
